backend/python/requirements.txt
openai-whisper
vaderSentiment
transformers
torch
pyttsx
ngrok
uvicorn
fastapi
ngrok
python-multipart
torchvision
transformers
huggingface_hub